Even more refactor?
