+++ src/main/java/ir/mctracker/api/PlayerVoteRewardReceiveEvent.java
@@ -1,4 +1,4 @@
-package ir.mctracker.mctrackervote.api;
+package ir.mctracker.api;
 import org.bukkit.event.Cancellable;
 import org.bukkit.event.Event;
 import org.bukkit.event.HandlerList;
