+++ src/main/java/ir/mctracker/core/config/Messages.java
@@ -1,8 +1,8 @@
-package ir.mctracker.mctrackervote.config;
+package ir.mctracker.core.config;
 
 import ir.jeykey.megacore.MegaPlugin;
 import ir.jeykey.megacore.config.Configurable;
-import ir.mctracker.mctrackervote.utilities.Util;
+import ir.mctracker.core.utilities.Util;
 
 import java.util.List;
 import java.util.Objects;
