+++ src/main/java/ir/mctracker/core/utilities/Metrics.java
@@ -1,4 +1,4 @@
-package ir.mctracker.mctrackervote.utilities;
+package ir.mctracker.core.utilities;
 
 import java.io.BufferedReader;
 import java.io.ByteArrayOutputStream;
